Added libraries to get icons and to handle the back button on compose.
Use adaptative libraries to create a layout such that displays a pane with the list of groups and when a group is selected, displays a pane with all locations that belong to that group. It automatically adapts to the screen dimensions to decide how many panes to show at once.